Address Omni launcher review comments
