+++ examples/launchers/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Model-local launcher presets for examples/run_omni.py."""
