+++ tests/unit_test/ming_omni/test_pipeline.py
@@ -9,6 +9,11 @@ import sys
 from pathlib import Path
 from types import ModuleType, SimpleNamespace
 
+from examples.launchers.ming_omni import (
+    launch_ming_speech_server as _launch_speech_server,
+)
+from examples.launchers.ming_omni import launch_ming_text_server as _launch_text_server
+
 
 def test_ming_text_config_imports_and_uses_current_stage_schema() -> None:
     from sglang_omni.models.ming_omni.config import MingOmniPipelineConfig
@@ -97,10 +102,6 @@ def test_ming_speech_launcher_exposes_tp_size_arg(monkeypatch) -> None:
 
 
 def test_ming_speech_launcher_places_thinker_tp_and_talker(monkeypatch) -> None:
-    from examples._omni_launcher import (
-        launch_ming_speech_server as _launch_speech_server,
-    )
-
     captured: dict[str, object] = {}
     serve_module = ModuleType("sglang_omni.serve")
 
@@ -226,8 +227,6 @@ def test_ming_audio_encoder_moves_inputs_to_component_device() -> None:
 
 
 def test_ming_text_launcher_places_tp_ranks_on_distinct_gpus(monkeypatch) -> None:
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
-
     captured: dict[str, object] = {}
     serve_module = ModuleType("sglang_omni.serve")
 
@@ -263,9 +262,42 @@ def test_ming_text_launcher_places_tp_ranks_on_distinct_gpus(monkeypatch) -> Non
     assert thinker.gpu == [0, 1, 2]
 
 
-def test_ming_text_launcher_allows_encoder_gpu_overrides(monkeypatch) -> None:
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
+def test_ming_text_launcher_rejects_nonpositive_tp_before_config(monkeypatch) -> None:
+    import pytest
 
+    from sglang_omni.models.ming_omni import config as config_module
+
+    def fail_config_build(*args, **kwargs):
+        raise AssertionError("config must not be built")
+
+    monkeypatch.setattr(config_module, "MingOmniPipelineConfig", fail_config_build)
+    serve_module = ModuleType("sglang_omni.serve")
+    serve_module.launch_server = lambda *args, **kwargs: None
+    monkeypatch.setitem(sys.modules, "sglang_omni.serve", serve_module)
+
+    for tp_size in (0, -1):
+        args = SimpleNamespace(
+            model_path="dummy",
+            relay_backend="shm",
+            tp_size=tp_size,
+            quantization=None,
+            cpu_offload_gb=0,
+            gpu_audio_encoder=None,
+            gpu_image_encoder=None,
+            image_encoder_tp=1,
+            thinker_only=False,
+            mem_fraction_static=None,
+            thinker_max_seq_len=8192,
+            host="127.0.0.1",
+            port=8000,
+            model_name="ming-omni",
+        )
+
+        with pytest.raises(ValueError, match="--tp-size must be >= 1"):
+            _launch_text_server(args)
+
+
+def test_ming_text_launcher_allows_encoder_gpu_overrides(monkeypatch) -> None:
     captured: dict[str, object] = {}
     serve_module = ModuleType("sglang_omni.serve")
 
@@ -305,8 +337,6 @@ def test_ming_text_launcher_allows_encoder_gpu_overrides(monkeypatch) -> None:
 def test_ming_text_launcher_can_build_thinker_only_smoke_pipeline(
     monkeypatch,
 ) -> None:
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
-
     captured: dict[str, object] = {}
     serve_module = ModuleType("sglang_omni.serve")
 
@@ -347,8 +377,6 @@ def test_ming_text_launcher_can_build_thinker_only_smoke_pipeline(
 
 
 def test_ming_text_launcher_configures_image_encoder_tp(monkeypatch) -> None:
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
-
     captured: dict[str, object] = {}
     serve_module = ModuleType("sglang_omni.serve")
 
@@ -387,8 +415,6 @@ def test_ming_text_launcher_configures_image_encoder_tp(monkeypatch) -> None:
 def test_ming_text_launcher_rejects_image_encoder_tp_zero(monkeypatch) -> None:
     import pytest
 
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
-
     serve_module = ModuleType("sglang_omni.serve")
     serve_module.launch_server = lambda *a, **kw: None
     monkeypatch.setitem(sys.modules, "sglang_omni.serve", serve_module)
@@ -418,8 +444,6 @@ def test_ming_text_launcher_rejects_thinker_only_with_image_encoder_tp(
     monkeypatch,
 ) -> None:
     import pytest
-
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
 
     serve_module = ModuleType("sglang_omni.serve")
     serve_module.launch_server = lambda *a, **kw: None
@@ -451,8 +475,6 @@ def test_ming_text_launcher_requires_gpu_ids_for_image_encoder_tp(
 ) -> None:
     import pytest
 
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
-
     serve_module = ModuleType("sglang_omni.serve")
     serve_module.launch_server = lambda *a, **kw: None
     monkeypatch.setitem(sys.modules, "sglang_omni.serve", serve_module)
@@ -481,8 +503,6 @@ def test_ming_text_launcher_requires_gpu_ids_for_image_encoder_tp(
 def test_ming_text_launcher_rejects_mismatched_gpu_count(monkeypatch) -> None:
     import pytest
 
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
-
     serve_module = ModuleType("sglang_omni.serve")
     serve_module.launch_server = lambda *a, **kw: None
     monkeypatch.setitem(sys.modules, "sglang_omni.serve", serve_module)
@@ -510,8 +530,6 @@ def test_ming_text_launcher_rejects_mismatched_gpu_count(monkeypatch) -> None:
 
 def test_ming_text_launcher_rejects_duplicate_gpu_ids(monkeypatch) -> None:
     import pytest
-
-    from examples._omni_launcher import launch_ming_text_server as _launch_text_server
 
     serve_module = ModuleType("sglang_omni.serve")
     serve_module.launch_server = lambda *a, **kw: None
